+++ _1327/settings.py
@@ -39,7 +39,6 @@ INSTALLED_APPS = (
     'static_precompiler',
     'bootstrap3',
     '_1327.main',
-    '_1327.auth',
 )
 
 MIDDLEWARE_CLASSES = (
